src/server/requirements.txt
Django==4.1.2
djangorestframework
django-filter
markdown
psycopg2
requests
grpcio
grpcio-tools
